refactoring, memory allocarion device operation by device
